backend/requirements.txt
aniso8601==9.0.1
appdirs==1.4.4
asgiref==3.7.2
baron==0.10.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
defusedxml==0.7.1
dj-database-url==2.0.0
dj-email-url==1.0.6
dj-rest-auth==4.0.1
Django==4.2.3
django-allauth==0.54.0
django-cache-url==3.4.4
django-ckeditor==6.7.0
django-cors-headers==4.2.0
django-environs==0.1.1
django-filter==23.2
django-js-asset==2.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
environs==9.5.0
graphene==3.3
graphene-django==3.1.3
graphql-core==3.2.3
graphql-relay==3.2.0
gunicorn==21.2.0
idna==3.4
marshmallow==3.20.1
oauthlib==3.2.2
packaging==23.1
Pillow==10.0.0
promise==2.3
psycopg2==2.9.6
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
redbaron==0.9.2
requests==2.31.0
requests-oauthlib==1.3.1
rply==0.7.8
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.7.1
urllib3==2.0.4
